+++ lib/ts/recipe/webauthn/recipeImplementation.ts
@@ -160,7 +160,13 @@ export default function getRecipeImplementation(
                 fetchResponse,
             };
         },
-        signUp: async function ({ webauthnGeneratedOptionsId, credential, options, userContext }) {
+        signUp: async function ({
+            webauthnGeneratedOptionsId,
+            credential,
+            shouldTryLinkingWithSessionUser,
+            options,
+            userContext,
+        }) {
             const { jsonBody, fetchResponse } = await querier.post<
                 | {
                       status: "OK";
@@ -185,6 +191,7 @@ export default function getRecipeImplementation(
                     body: JSON.stringify({
                         webauthnGeneratedOptionsId,
                         credential,
+                        shouldTryLinkingWithSessionUser,
                     }),
                 },
                 Querier.preparePreAPIHook({
@@ -205,7 +212,13 @@ export default function getRecipeImplementation(
                 fetchResponse,
             };
         },
-        signIn: async function ({ webauthnGeneratedOptionsId, credential, options, userContext }) {
+        signIn: async function ({
+            webauthnGeneratedOptionsId,
+            credential,
+            shouldTryLinkingWithSessionUser,
+            options,
+            userContext,
+        }) {
             const { jsonBody, fetchResponse } = await querier.post<
                 | {
                       status: "OK";
@@ -226,6 +239,7 @@ export default function getRecipeImplementation(
                     body: JSON.stringify({
                         webauthnGeneratedOptionsId,
                         credential,
+                        shouldTryLinkingWithSessionUser,
                     }),
                 },
                 Querier.preparePreAPIHook({
@@ -383,7 +397,12 @@ export default function getRecipeImplementation(
                 registrationResponse,
             };
         },
-        registerCredentialWithSignUp: async function ({ email, options, userContext }) {
+        registerCredentialWithSignUp: async function ({
+            email,
+            shouldTryLinkingWithSessionUser,
+            options,
+            userContext,
+        }) {
             // Get the registration options by using the passed email ID.
             const registrationOptions = await this.getRegisterOptions({ options, userContext, email });
             if (registrationOptions?.status !== "OK") {
@@ -410,6 +429,7 @@ export default function getRecipeImplementation(
             return await this.signUp({
                 webauthnGeneratedOptionsId: registrationOptions.webauthnGeneratedOptionsId,
                 credential: registerCredentialResponse.registrationResponse,
+                shouldTryLinkingWithSessionUser,
                 options,
                 userContext,
             });
@@ -437,7 +457,7 @@ export default function getRecipeImplementation(
                 authenticationResponse: authenticationResponse,
             };
         },
-        authenticateCredentialWithSignIn: async function ({ options, userContext }) {
+        authenticateCredentialWithSignIn: async function ({ shouldTryLinkingWithSessionUser, options, userContext }) {
             // Make a call to get the sign in options using the entered email ID.
             const signInOptions = await this.getSignInOptions({ options, userContext });
             if (signInOptions?.status !== "OK") {
@@ -459,6 +479,7 @@ export default function getRecipeImplementation(
             return await this.signIn({
                 webauthnGeneratedOptionsId: signInOptions.webauthnGeneratedOptionsId,
                 credential: authenticateCredentialResponse.authenticationResponse,
+                shouldTryLinkingWithSessionUser,
                 options: options,
                 userContext: userContext,
             });